sam normalized and RGB uint8

+++ src/vollseg/utils.py
@@ -4541,7 +4541,7 @@ def VollSamZ(
         channel_image = channel_image * 255
 
         instance_labels_currentz = mask_generator.generate(
-            channel_image.astype(np.float16)
+            channel_image.astype(np.uint8)
         )
         instance_labels_currentz = remove_small_objects(
             instance_labels_currentz.astype("uint16"), min_size=min_size
